[NewsingIT Project 2016/09/12 02:42] - FollowingList / FollowerList / TagSearch / Iamge buf fix / FCM / AlarmList

+++ app/src/main/java/com/example/apple/newsingit_project/data/json_data/followinglist/FollowingListRequestResults.java
@@ -2,11 +2,10 @@ package com.example.apple.newsingit_project.data.json_data.followinglist;
 
 public class FollowingListRequestResults {
     private String name;
-    private long id;
+    private int id;
     private String pf_url;
     private String aboutme;
     private boolean flag;
-
 
     public boolean getFlag() {
         return flag;
@@ -24,11 +23,11 @@ public class FollowingListRequestResults {
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return this.id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
